+++ src/tosvg/org/girod/javafx/svgimage/tosvg/converters/GroupConverter.java
@@ -40,7 +40,7 @@ import org.girod.javafx.svgimage.tosvg.xml.XMLNode;
  * A converter which convert Groups.
  *
  * @since 1.0
-   */
+ */
 public class GroupConverter extends AbstractConverter {
    private Group group = null;
 
@@ -71,8 +71,10 @@ public class GroupConverter extends AbstractConverter {
       if (clipID != null) {
          StringBuilder buf = new StringBuilder();
          String style = buf.toString();
-         node.addAttribute("style", style);
-      }      
+         if (!style.isEmpty()) {
+            node.addAttribute("style", style);
+         }
+      }
    }
 
    /**